FROMLIST: mach-k3: use common MMU table for all k3
k3_mem_map is used by u-boot to configure the page tables for the MMU as
part of enable_cache. Currently we provide a common mmu table upstream
and a hack table for j722s downstream but this solution does not scale
across platforms with different memory configurations (like 62x SiP).

Therefore this patch replaces the two existing MMU tables with a single
minimal table that could be modified at runtime as per the device-tree.

The 128MiB region starting from SPL_TEXT_BASE is mapped for use by the
A-core SPL. 128MiB provides enough space for future use cases such as
falcon boot from the A-Core SPL etc.

Signed-off-by: Anshul Dalal <anshuld@ti.com>